% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{assignPointsToLines}
\alias{assignPointsToLines}
\title{Assign Points To Lines}
\usage{
assignPointsToLines(points, lines, maxDist = NA)
}
\arguments{
\item{points}{spatial points}

\item{lines}{spatial lines}

\item{maxDist}{Boolean Flag}
}
\value{
spatial line df
}
\description{
Assign Points To Lines
}
\examples{
\dontrun{
assignPointsToLines(points, lines, maxDist)
}
}
\seealso{
Other nwasc.internals: \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
