% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_output.R
\name{bind_phases}
\alias{bind_phases}
\title{Combine Phase 1 DTS & CTS segmented data with Phase 2 CTS}
\usage{
bind_phases(seg.dat.phase2, seg.dat.phase1.cts, seg.dat.phase1.dts)
}
\arguments{
\item{seg.dat.phase2}{df segmented phase 2 data}

\item{seg.dat.phase1.cts}{df segment phase 1 cts}

\item{seg.dat.phase1.dts}{df segmented dts data from phase 1}
}
\value{
df wide form segmented data from NWASC with columns of species counts
}
\description{
Combines the three pieces of the segment catalog and replaces missing value observatins
with a 0 count
}
\examples{
segmented_seabird_catalogs <- bind_phases(seg.dat.phase2, seg.dat.phase1.cts, seg.dat.phase1.dts)

}
