% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{calculate_seg_midpoints}
\alias{calculate_seg_midpoints}
\title{Calculate Segment Midpoints}
\usage{
calculate_seg_midpoints(lineframe, projHOM)
}
\arguments{
\item{lineframe}{spatial line object}

\item{projHOM}{projection string}
}
\value{
df of segment midpoints
}
\description{
Calculate Segment Midpoints
}
\examples{
midpoints <- calculate_seg_midpoints(lineframe,projHOM)
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
