% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{calculate_segment_endpoints}
\alias{calculate_segment_endpoints}
\title{Calculate Segment Endpoints}
\usage{
calculate_segment_endpoints(seg.all)
}
\arguments{
\item{seg.all}{df of all segmetns}
}
\value{
df location of segment endpoints
}
\description{
Calculates the endpoints of each segment
}
\examples{
end.pts <- calculate_segment_endpoints(seg.all)
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
