% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{create_all_segments}
\alias{create_all_segments}
\title{Create all segments}
\usage{
create_all_segments(seg, seg.empty, seg.length)
}
\arguments{
\item{seg}{df segmented tracks}

\item{seg.empty}{df of empty segments}

\item{seg.length}{numeric desired segment length}
}
\value{
df of all segments
}
\description{
Combines non-empty and empty segments
}
\examples{
\dontrun{
seg.all <- create_all_segments(seg, seg.empty, seg.length)
}
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
