% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{create_empty_segments}
\alias{create_empty_segments}
\title{Create Empty Segments}
\usage{
create_empty_segments(seg)
}
\arguments{
\item{seg}{segmented tracks}
}
\value{
df of empty segments
}
\description{
Create Empty Segments
}
\examples{
\dontrun{
seg.empty <- create_empty_segments(seg)
}
}
\seealso{
\code{\link{segment_tracks}}

Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
