% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{create_segmented_df}
\alias{create_segmented_df}
\title{Create Segmented Dataframe}
\usage{
create_segmented_df(seg.mids, seg.obs, transects)
}
\arguments{
\item{seg.mids}{df of segments with midpoint coordinates}

\item{seg.obs}{df of segments with  observations}

\item{transects}{df of transect info from transect table}
}
\value{
df segmented nwasc data with midpoint and observations
}
\description{
Combine midpoints, summarizes observations and filter non-observations
}
\examples{
segmented <- create_segmented_df(seg.mids, seg.obs, transects)
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
