% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{define_lineframe}
\alias{define_lineframe}
\title{Define Lineframe}
\usage{
define_lineframe(linelist, projHOM)
}
\arguments{
\item{linelist}{spatail Line object}

\item{projHOM}{projecton}
}
\value{
spatial dataframe
}
\description{
Define Lineframe
}
\examples{
lineframe <- define_lineframe(linelist, projHOM)
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
