% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{listLines}
\alias{listLines}
\title{Create Spatial Lines}
\usage{
listLines(df)
}
\arguments{
\item{df}{grouped df of segments}
}
\value{
spatial Line object
}
\description{
Create Spatial Lines
}
\examples{
\dontrun{
linelist <- seg.all.new \%>\%
group_by(transect_id, id) \%>\%
  do(coords = listLines(.))
}
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{obs2Lines}}, \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
