% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwasc-package.R
\docType{package}
\name{nwasc-package}
\alias{nwasc}
\alias{nwasc-package}
\title{nwasc: What the Package Does (Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The nwasc package was written by FWS for BOEM to segment transect 
  survey data from the Northwest Atlantic Seabird Catalog (maintained by FWS). 
  Each transect is segmented into fixed length segments with the remaineder 
  placed according to a set rule. Observations are tallied by segment and 
  transect.
}
\author{
\strong{Maintainer}: Robert Fowler \email{robert_fowler@fws.gov}

}
\keyword{internal}
