% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwasc.ph1.dts.dat}
\alias{nwasc.ph1.dts.dat}
\title{Legacy NWASC Phase 1 DTS Transect Data}
\format{A data frame with 37818 rows and 17 variables}
\source{
\url{https://github.com/USFWS/AMAPPS/tree/master/NWASC}
}
\usage{
nwasc.ph1.dts.dat
}
\description{
Historic DTS data from NWASC
}
\keyword{datasets}
