% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwasc.ph1.dts.obs.dat}
\alias{nwasc.ph1.dts.obs.dat}
\title{Legacy NWASC Phase 1 DTS Observation Data}
\format{A data frame with 372817 rows and 8 varibles}
\source{
\url{https://github.com/USFWS/AMAPPS/tree/master/NWASC}
}
\usage{
nwasc.ph1.dts.obs.dat
}
\description{
Historic DTS data from NWASC
}
\keyword{datasets}
