% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwasc.ph1.shp.pre}
\alias{nwasc.ph1.shp.pre}
\title{Example of NWASC Phase 1 CTS Track Data}
\format{A data frame with 2837748 rows and 6 variables:
\describe{
\item{start_dt}{Start date}
\item{survey_method_cd}{Methodology of survey}
\item{survey_type_cd}{Survey type code}
\item{source_dataset_id}{Id of source dataset}
\item{transect_id}{Id of transecet}
\item{segmented_transect_id}{source_transect_id from catalog}
\item{transect_width_nb}{width of transect when availiable}
}}
\source{
\url{https://github.com/USFWS/AMAPPS/tree/master/NWASC}
}
\usage{
nwasc.ph1.shp.pre
}
\description{
A daset example of a cleaned up query of phase 1 track information
from the Northwest Atlantic Seabird Catalog
}
\keyword{datasets}
