% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwasc.ph2.obs.pre}
\alias{nwasc.ph2.obs.pre}
\title{Example of NWASC Phase 1 CTS Observation Data}
\format{A data frame with 90531 rows and 5 variables:
\describe{
\item{long}{longitude}
\item{lat}{latitude}
\item{transect_id}{Id of transect}
\item{spp_cd}{species code}
\item{count}{observation count}
}}
\source{
\url{https://github.com/USFWS/AMAPPS/tree/master/NWASC}
}
\usage{
nwasc.ph2.obs.pre
}
\description{
A dataset example of a cleaned up query of phase 1 observation information
from the Northwest Atlantic Seabird Catalog
}
\keyword{datasets}
