% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwasc.ph2.shp.pre}
\alias{nwasc.ph2.shp.pre}
\title{Example of NWASC Phase 1 CTS Track Data}
\format{A data frame with 526991 rows and 6 variables:
\describe{
\item{long}{longitude}
\item{lat}{latitude}
\item{dataset_id}{Id of source dataset}
\item{piece}{Transect piece ID}
\item{order}{Index of order}
\item{transect_id}{Id of transect}
}}
\source{
\url{https://github.com/USFWS/AMAPPS/tree/master/NWASC}
}
\usage{
nwasc.ph2.shp.pre
}
\description{
A dataset example of a cleaned up query of phase 1 track information
from the Northwest Atlantic Seabird Catalog
}
\keyword{datasets}
