% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwasc-package.R
\docType{package}
\name{nwasc}
\alias{nwasc}
\alias{nwasc-package}
\title{nwasc: Segmentation of the Northwest Atlantic Seabird Catalog}
\description{
The nwasc package segments the dts and cts surveys in the Northwest Atlantic
Seabird Catalog (NWASC).
}
\section{nwasc functions}{

The nwasc functions: all exported functions begin with "nwasc_"

nwasc_SegmentCTS: Segments CTS type data returning a wide format

nwasc_SegmentDTS: Segments DTS type data returning a wide format

nwasc_bind_phases: Binds phase 1 and phase 2 data and fixes NA counts

nwasc_write_segmented_csv: writes a date stamped .csv of the segmented data
}

\author{
\strong{Maintainer}: Robert Fowler \email{robert_fowler@fws.gov}

}
\keyword{internal}
