% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_output.R
\name{nwasc_write_segmented_csv}
\alias{nwasc_write_segmented_csv}
\title{Save segmentation results to a .csv file}
\usage{
nwasc_write_segmented_csv(segmented_seabird_catalogs)
}
\arguments{
\item{segmented_seabird_catalogs}{segmented nwasc dataframe}
}
\value{
csv with date and draft name
}
\description{
Save segmentation results to a .csv file
}
\section{WARNING!}{
 this may be a very large file.  zipping and removing is advised
}

\examples{
\dontrun{
write_segmented_csv(segmented_seabird_catalogs)
}
}
