% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{obs2Lines}
\alias{obs2Lines}
\title{Place Observations on Segment Line}
\usage{
obs2Lines(df, lineframe, maxDist, projHOM)
}
\arguments{
\item{df}{df of filtered observations}

\item{lineframe}{spatial object of segments}

\item{maxDist}{maximum distance a observation is allowed from a segment NA ignores this feature}
}
\value{
spatial line df
}
\description{
Place Observations on Segment Line
}
\examples{
seg.obs <- observations \%>\%
filter(transect_id \%in\% seg.mids$transect_id) \%>\%
  group_by(transect_id) \%>\%
    do(obs2Lines(., lineframe, maxDist, projHOM))
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{place_midpoints}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
