% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{place_midpoints}
\alias{place_midpoints}
\title{Place Midpoints}
\usage{
place_midpoints(seg.all.new, midpoints, transects)
}
\arguments{
\item{seg.all.new}{df of segments}

\item{midpoints}{df of midpoints}

\item{transects}{df of transect info}
}
\value{
df of segments with midpoint coordinates
}
\description{
Place Midpoints
}
\examples{
seg.mids <- place_midpoints(seg.all.new, midpoints, transects)
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_observations}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
