% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_utils.R
\name{place_observations}
\alias{place_observations}
\title{Place Observations}
\usage{
place_observations(observations, seg.mids, lineframe, maxDist, projHOM)
}
\arguments{
\item{observations}{df of observations}

\item{seg.mids}{df of segments with midopoints}

\item{lineframe}{spatial dataframe}

\item{maxDist}{maximum distance a observation is allowed from a segment NA ignores this feature}
}
\value{
df of segments with  observations
}
\description{
Place Observations
}
\examples{
seg.obs <- place_observations(observations, seg.mids, lineframe, maxDist, projHOM)
}
\seealso{
Other nwasc.internals: \code{\link{assignPointsToLines}},
  \code{\link{calculate_seg_midpoints}},
  \code{\link{calculate_segment_endpoints}},
  \code{\link{complete_segments}},
  \code{\link{create_all_segments}},
  \code{\link{create_empty_segments}},
  \code{\link{create_segmented_df}},
  \code{\link{define_lineframe}},
  \code{\link{duplicate_endpoints}},
  \code{\link{listLines}}, \code{\link{obs2Lines}},
  \code{\link{place_midpoints}},
  \code{\link{segment_tracks}}
}
\concept{nwasc.internals}
