% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cts_segmentation.R
\name{segmentCTS}
\alias{segmentCTS}
\title{Segment CTS data from the NWASC}
\usage{
segmentCTS(observations, tracks, transects, seg.length = 4,
  seg.tol = 0.5, seg.min = seg.length * seg.tol, maxDist = NA)
}
\arguments{
\item{observations}{df of observations from observeration table}

\item{tracks}{df of observations from tracks table (legacy translated from shapefiles)}

\item{transects}{df of transect info from transect table}

\item{seg.length}{numeric desired segment length}

\item{seg.tol}{numeric minimal proption of \code{seg.length} to be considered a segment}

\item{seg.min}{calculated shortest length of a valid segmetn}

\item{maxDist}{maximum distance a observation is allowed from a segment NA ignores this feature}
}
\value{
df wide format segment cts data
}
\description{
This is a wrapper function that takes queries from the observations, tracks and transects.
}
\section{Note on names}{

The following columns are need from each of the tables.
transects:
"start_dt", "dataset_id", "source_dataset_id",
"segmented_transect_id" (renamed from source_transect_id in queury!)
"transect_id", "transect_width_nb", "survey_type_cd", "survey_method_cd"
tracks: (this historically was also created from shapefiles example in rew-data/)
"lat", "long", "transect_id", "dataset_id, "piece", "order".
observations:
"transect_id", "lat", "long", "spp_cd", "count"
}

\section{WARNING}{
 double check names,
with special attention to renaming source_transect_id as "segmented_transect_id"
}

\examples{
seg.dat.phase2 = segmentCTS(nwasc.ph2.obs.pre,
                            nwasc.ph2.shp.pre,
                            nwasc.ph2.cts.dat,
                            seg.min = 0.0)

seg.dat.phase1.cts = segmentCTS(nwasc.ph1.obs.pre,
                                nwasc.ph1.shp.pre,
                                nwasc.ph1.cts.dat,
                                seg.min = 0.0)
}
