% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Legacy-dts-segmentation.R
\name{segmentDTS}
\alias{segmentDTS}
\title{Process dts survey data in NWASC database}
\usage{
segmentDTS(observations, transects, v.spd = 10, occurences = FALSE)
}
\arguments{
\item{observations}{df of observations from phase 1 DTS surveys}

\item{transects}{df of transect information from phase 1 DTS surveys}

\item{v.spd}{numeric approx. speed of vessel}

\item{occurences}{boolean flag  Calculates species counts by default (FALSE);
set occurences = TRUE for number of flock sightings}
}
\value{
df wide-form dataframe with processed data
}
\description{
Process dts survey data in NWASC database
Returns wide-form dataframe with processed data
}
\details{
Requires observation table with lat/long for each sighting and transect table with single row for each transect id

v.spd sets assumed vessel speed in knots when vessel speed is missing
Calculates species counts by default; set occurences = TRUE for number of flock sightings

Distances are in nautical miles

Kyle Dettloff
Modified 09-30-16

read in dts and observation tables
}
\examples{
seg.dat.phase1.dts = segmentDTS(nwasc.ph1.dts.obs.dat, nwasc.ph1.dts.dat)
}
