% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwasc-package.R
\docType{package}
\name{nwasc-package}
\alias{nwasc}
\alias{nwasc-package}
\title{nwasc: What the Package Does (Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
