% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwasc.ph2.shp.pre}
\alias{nwasc.ph2.shp.pre}
\title{Example of NWASC Phase 1 CTS Track Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 526991 rows and 6 columns.}
\source{
\url{https://github.com/USFWS/AMAPPS/tree/master/NWASC}
}
\usage{
nwasc.ph2.shp.pre
}
\description{
A dataset example of a cleaned up query of phase 1 track information
from the Northwest Atlantic Seabird Catalog
}
\keyword{datasets}
